+++ src/lib.rs
@@ -33,6 +33,7 @@ pub mod safe_list;
 pub mod scanner;
 pub mod scanner_pool;
 pub mod serde_utils;
+pub mod slack;
 pub mod snippet;
 pub mod update;
 pub mod util;