+++ tests/int_gitlab.rs
@@ -68,6 +68,8 @@ fn test_gitlab_remote_scan() -> Result<()> {
             jira_url: None,
             jql: None,
             max_results: 100,
+            slack_query: None,
+            slack_api_url: Url::parse("https://slack.com/api/").unwrap(),
             // Docker image scanning
             docker_image: Vec::new(),
             git_clone: GitCloneMode::Bare,