+++ tests/int_vulnerable_files.rs
@@ -68,6 +68,8 @@ impl TestContext {
                 jira_url: None,
                 jql: None,
                 max_results: 100,
+                slack_query: None,
+                slack_api_url: Url::parse("https://slack.com/api/").unwrap(),
                 // Docker image scanning
                 docker_image: Vec::new(),
                 // git clone / history options
@@ -138,6 +140,8 @@ impl TestContext {
                 jira_url: None,
                 jql: None,
                 max_results: 100,
+                slack_query: None,
+                slack_api_url: Url::parse("https://slack.com/api/").unwrap(),
                 // Docker image scanning
                 docker_image: Vec::new(),
                 // git clone / history options