+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.28.0]
+- Added support for scanning Slack
+
 ## [1.27.0]
 - Added Buildkite rule
 - Added support for scanning Docker images via `--docker-image`